Threading/Scheduler rewrite, UsbHid fix, pci enumeration fix
So OUT OF THE BLUE a scheduler error appeared, and it was a sort of "why
does nothing work now" bug. So I decided to review my whole threading
and scheduler and it dawned upon me ... "What the FRICK did I do to my
scheduler?!" Anyway rewrote much of the system to work with seperate
lists, seperate systems instead of being interlinked as they were.
Everything works much better now.

The scheduler has a weakness though, it needs new list-nodes EACH time
it schedules a task, which is serious shit, to counter-act this problem
I made a node-recycler in the scheduler, but at some point this probably
will have to be fixed for performance.

+++ modules/Usb/UsbCore/Devices/UsbHid.c
@@ -164,7 +164,7 @@ void UsbHidInit(UsbHcDevice_t *UsbDevice, uint32_t InterfaceIndex)
 	if (UsbFunctionGetDescriptor((UsbHc_t*)UsbDevice->HcDriver, UsbDevice->Port,
 		ReportDescriptor, USB_REQUEST_DIR_IN | USB_REQUEST_TARGET_INTERFACE,
 		HidDescriptor->ClassDescriptorType,
-		0, InterfaceIndex, HidDescriptor->ClassDescriptorLength) != TransferFinished)
+		0, (uint8_t)InterfaceIndex, HidDescriptor->ClassDescriptorLength) != TransferFinished)
 	{
 		LogFatal("USBH", "Failed to get Report Descriptor.");
 		kfree(ReportDescriptor);
